+++ src/types/bytes.rs
@@ -1,3 +1,4 @@
+use crate::inspect::types::TypeInfo;
 use crate::{
     ffi, AsPyPointer, FromPyObject, IntoPy, Py, PyAny, PyObject, PyResult, PyTryFrom, Python,
     ToPyObject,
@@ -128,13 +129,22 @@ impl<'a> IntoPy<PyObject> for &'a [u8] {
     fn into_py(self, py: Python<'_>) -> PyObject {
         PyBytes::new(py, self).to_object(py)
     }
+
+    fn type_output() -> TypeInfo {
+        TypeInfo::builtin("bytes")
+    }
 }
 
 impl<'a> FromPyObject<'a> for &'a [u8] {
     fn extract(obj: &'a PyAny) -> PyResult<Self> {
         Ok(<PyBytes as PyTryFrom>::try_from(obj)?.as_bytes())
     }
+
+    fn type_input() -> TypeInfo {
+        Self::type_output()
+    }
 }
+
 #[cfg(test)]
 mod tests {
     use super::PyBytes;